+++ NEWS.md
@@ -1,5 +1,7 @@
 # tidycomm 0.1.0.9000
 
+## New features
+* Added Fretwurst's Lotus and S-Lotus to test_icr function
 
 ## Bugfixes
 * Empty groups are dropped if describing with more than one grouping variable